Split extension pbrSpecularGlossiness blender into own file
