+++ addons/io_scene_gltf2/blender/imp/gltf2_blender_KHR_materials_pbrSpecularGlossiness.py
@@ -18,80 +18,24 @@
  * Contributor(s): Julien Duroure.
  *
  * ***** END GPL LICENSE BLOCK *****
- * This development is done in strong collaboration with Airbus Defence & Space
  """
 
-from .....io.com.gltf2_io_texture import *
 import bpy
+from .gltf2_blender_texture import *
 
-class KHR_materials_pbrSpecularGlossiness():
-
-    SIMPLE  = 1
-    TEXTURE = 2
-    TEXTURE_FACTOR = 3
-
-    def __init__(self, json, gltf):
-        self.json = json # KHR_materials_pbrSpecularGlossiness json
-        self.gltf = gltf # Reference to global glTF instance
-
-        self.diffuse_type   = self.SIMPLE
-        self.specgloss_type = self.SIMPLE
-        self.vertex_color   = False
-
-        # Default Values
-        self.diffuseFactor    = [1.0,1.0,1.0,1.0]
-        self.glossinessFactor = 1.0
-        self.specularFactor   = [1.0,1.0,1.0]
-
-    def read(self):
-        if self.json is None:
-            return # will use default values
-
-        if 'diffuseTexture' in self.json.keys():
-            self.diffuse_type = self.TEXTURE
-            self.diffuseTexture = PyTexture(self.json['diffuseTexture']['index'], self.gltf.json['textures'][self.json['diffuseTexture']['index']], self.gltf)
-            self.diffuseTexture.read()
-
-            if 'texCoord' in self.json['diffuseTexture']:
-                self.diffuseTexture.texcoord = int(self.json['diffuseTexture']['texCoord'])
-            else:
-                self.diffuseTexture.texcoord = 0
-
-        if 'diffuseFactor' in self.json.keys():
-            self.diffuseFactor = self.json['diffuseFactor']
-            if self.diffuse_type == self.TEXTURE and self.diffuseFactor != [1.0,1.0,1.0,1.0]:
-                self.diffuse_type = self.TEXTURE_FACTOR
-
-        if 'specularGlossinessTexture' in self.json.keys():
-            self.specgloss_type = self.TEXTURE
-            self.specularGlossinessTexture = PyTexture(self.json['specularGlossinessTexture']['index'], self.gltf.json['textures'][self.json['specularGlossinessTexture']['index']], self.gltf)
-            self.specularGlossinessTexture.read()
-
-            if 'texCoord' in self.json['specularGlossinessTexture']:
-                self.specularGlossinessTexture.texcoord = int(self.json['specularGlossinessTexture']['texCoord'])
-            else:
-                self.specularGlossinessTexture.texcoord = 0
-
-        if 'glossinessFactor' in self.json.keys():
-            self.glossinessFactor = self.json['glossinessFactor']
-
-        if 'specularFactor' in self.json.keys():
-            self.specularFactor = self.json['specularFactor']
-            if self.specgloss_type == self.TEXTURE and self.specgloss_type != [1.0,1.0,1.0]:
-                self.specgloss_type = self.TEXTURE_FACTOR
+class BlenderKHR_materials_pbrSpecularGlossiness():
 
 
-    def use_vertex_color(self):
-        self.vertex_color = True
-
-    def create_blender(self, mat_name):
+    @staticmethod
+    def create(pyext, mat_name):
         engine = bpy.context.scene.render.engine
         if engine == 'CYCLES':
-            self.create_blender_cycles(mat_name)
+            BlenderKHR_materials_pbrSpecularGlossiness.create_cycles(pyext, mat_name)
         else:
             pass #TODO for internal / Eevee in future 2.8
 
-    def create_blender_cycles(self, mat_name):
+    @staticmethod
+    def create_cycles(pyext, mat_name):
         material = bpy.data.materials[mat_name]
         material.use_nodes = True
         node_tree = material.node_tree
@@ -112,12 +56,12 @@ class KHR_materials_pbrSpecularGlossiness():
         mix        = node_tree.nodes.new('ShaderNodeMixShader')
         mix.location     = 500,0
 
-        glossy.inputs[1].default_value = 1 - self.glossinessFactor
+        glossy.inputs[1].default_value = 1 - pyext.glossinessFactor
 
-        if self.diffuse_type == self.SIMPLE:
-            if not self.vertex_color:
+        if pyext.diffuse_type == pyext.SIMPLE:
+            if not pyext.vertex_color:
                 # change input values
-                diffuse.inputs[0].default_value = self.diffuseFactor
+                diffuse.inputs[0].default_value = pyext.diffuseFactor
 
             else:
                 # Create attribute node to get COLOR_0 data
@@ -128,10 +72,10 @@ class KHR_materials_pbrSpecularGlossiness():
                 # links
                 node_tree.links.new(diffuse.inputs[0], attribute_node.outputs[1])
 
-        elif self.diffuse_type == self.TEXTURE_FACTOR:
+        elif pyext.diffuse_type == pyext.TEXTURE_FACTOR:
 
             #TODO alpha ?
-            if self.vertex_color:
+            if pyext.vertex_color:
                 # TODO tree locations
                 # Create attribute / separate / math nodes
                 attribute_node = node_tree.nodes.new('ShaderNodeAttribute')
@@ -147,11 +91,11 @@ class KHR_materials_pbrSpecularGlossiness():
                 math_vc_B = node_tree.nodes.new('ShaderNodeMath')
                 math_vc_B.operation = 'MULTIPLY'
 
-            self.diffuseTexture.blender_create()
+            BlenderTexture.create(pyext.diffuseTexture)
 
             # create UV Map / Mapping / Texture nodes / separate & math and combine
             text_node = node_tree.nodes.new('ShaderNodeTexImage')
-            text_node.image = bpy.data.images[self.diffuseTexture.image.blender_image_name]
+            text_node.image = bpy.data.images[pyext.diffuseTexture.image.blender_image_name]
             text_node.location = -1000,500
 
             combine = node_tree.nodes.new('ShaderNodeCombineRGB')
@@ -160,17 +104,17 @@ class KHR_materials_pbrSpecularGlossiness():
             math_R  = node_tree.nodes.new('ShaderNodeMath')
             math_R.location = -500, 750
             math_R.operation = 'MULTIPLY'
-            math_R.inputs[1].default_value = self.diffuseFactor[0]
+            math_R.inputs[1].default_value = pyext.diffuseFactor[0]
 
             math_G  = node_tree.nodes.new('ShaderNodeMath')
             math_G.location = -500, 500
             math_G.operation = 'MULTIPLY'
-            math_G.inputs[1].default_value = self.diffuseFactor[1]
+            math_G.inputs[1].default_value = pyext.diffuseFactor[1]
 
             math_B  = node_tree.nodes.new('ShaderNodeMath')
             math_B.location = -500, 250
             math_B.operation = 'MULTIPLY'
-            math_B.inputs[1].default_value = self.diffuseFactor[2]
+            math_B.inputs[1].default_value = pyext.diffuseFactor[2]
 
             separate = node_tree.nodes.new('ShaderNodeSeparateRGB')
             separate.location = -750, 500
@@ -180,11 +124,11 @@ class KHR_materials_pbrSpecularGlossiness():
 
             uvmap = node_tree.nodes.new('ShaderNodeUVMap')
             uvmap.location = -2000, 500
-            uvmap["gltf2_texcoord"] = self.diffuseTexture.texcoord # Set custom flag to retrieve TexCoord
+            uvmap["gltf2_texcoord"] = pyext.diffuseTexture.texcoord # Set custom flag to retrieve TexCoord
             # UV Map will be set after object/UVMap creation
 
             # Create links
-            if self.vertex_color:
+            if pyext.vertex_color:
                 node_tree.links.new(separate_vertex_color.inputs[0], attribute_node.outputs[0])
                 node_tree.links.new(math_vc_R.inputs[1], separate_vertex_color.outputs[0])
                 node_tree.links.new(math_vc_G.inputs[1], separate_vertex_color.outputs[1])
@@ -213,12 +157,12 @@ class KHR_materials_pbrSpecularGlossiness():
 
             node_tree.links.new(diffuse.inputs[0], combine.outputs[0])
 
-        elif self.diffuse_type == self.TEXTURE:
+        elif pyext.diffuse_type == pyext.TEXTURE:
 
-            self.diffuseTexture.blender_create()
+            BlenderTexture.create(pyext.diffuseTexture)
 
             #TODO alpha ?
-            if self.vertex_color:
+            if pyext.vertex_color:
                 # Create attribute / separate / math nodes
                 attribute_node = node_tree.nodes.new('ShaderNodeAttribute')
                 attribute_node.attribute_name = 'COLOR_0'
@@ -248,28 +192,28 @@ class KHR_materials_pbrSpecularGlossiness():
 
             # create UV Map / Mapping / Texture nodes / separate & math and combine
             text_node = node_tree.nodes.new('ShaderNodeTexImage')
-            text_node.image = bpy.data.images[self.diffuseTexture.image.blender_image_name]
-            if self.vertex_color:
+            text_node.image = bpy.data.images[pyext.diffuseTexture.image.blender_image_name]
+            if pyext.vertex_color:
                 text_node.location = -2000,500
             else:
                 text_node.location = -500,500
 
             mapping = node_tree.nodes.new('ShaderNodeMapping')
-            if self.vertex_color:
+            if pyext.vertex_color:
                 mapping.location = -2500,500
             else:
                 mapping.location = -1500,500
 
             uvmap = node_tree.nodes.new('ShaderNodeUVMap')
-            if self.vertex_color:
+            if pyext.vertex_color:
                 uvmap.location = -3000,500
             else:
                 uvmap.location = -2000,500
-            uvmap["gltf2_texcoord"] = self.diffuseTexture.texcoord # Set custom flag to retrieve TexCoord
+            uvmap["gltf2_texcoord"] = pyext.diffuseTexture.texcoord # Set custom flag to retrieve TexCoord
             # UV Map will be set after object/UVMap creation
 
             # Create links
-            if self.vertex_color:
+            if pyext.vertex_color:
                 node_tree.links.new(separate_vertex_color.inputs[0], attribute_node.outputs[0])
 
                 node_tree.links.new(math_vc_R.inputs[1], separate_vertex_color.outputs[0])
@@ -297,20 +241,20 @@ class KHR_materials_pbrSpecularGlossiness():
             node_tree.links.new(text_node.inputs[0], mapping.outputs[0])
 
 
-        if self.specgloss_type == self.SIMPLE:
+        if pyext.specgloss_type == pyext.SIMPLE:
 
             combine = node_tree.nodes.new('ShaderNodeCombineRGB')
-            combine.inputs[0].default_value = self.specularFactor[0]
-            combine.inputs[1].default_value = self.specularFactor[1]
-            combine.inputs[2].default_value = self.specularFactor[2]
+            combine.inputs[0].default_value = pyext.specularFactor[0]
+            combine.inputs[1].default_value = pyext.specularFactor[1]
+            combine.inputs[2].default_value = pyext.specularFactor[2]
 
             # links
             node_tree.links.new(glossy.inputs[0], combine.outputs[0])
 
-        elif self.specgloss_type == self.TEXTURE:
-            self.specularGlossinessTexture.blender_create()
+        elif pyext.specgloss_type == pyext.TEXTURE:
+            BlenderTexture.create(pyext.specularGlossinessTexture)
             spec_text = node_tree.nodes.new('ShaderNodeTexImage')
-            spec_text.image = bpy.data.images[self.specularGlossinessTexture.image.blender_image_name]
+            spec_text.image = bpy.data.images[pyext.specularGlossinessTexture.image.blender_image_name]
             spec_text.color_space = 'NONE'
             spec_text.location = -500,0
 
@@ -319,7 +263,7 @@ class KHR_materials_pbrSpecularGlossiness():
 
             spec_uvmap = node_tree.nodes.new('ShaderNodeUVMap')
             spec_uvmap.location = -1500,0
-            spec_uvmap["gltf2_texcoord"] = self.specularGlossinessTexture.texcoord # Set custom flag to retrieve TexCoord
+            spec_uvmap["gltf2_texcoord"] = pyext.specularGlossinessTexture.texcoord # Set custom flag to retrieve TexCoord
 
             # links
             node_tree.links.new(glossy.inputs[0], spec_text.outputs[0])
@@ -328,18 +272,18 @@ class KHR_materials_pbrSpecularGlossiness():
             node_tree.links.new(spec_mapping.inputs[0], spec_uvmap.outputs[0])
             node_tree.links.new(spec_text.inputs[0], spec_mapping.outputs[0])
 
-        elif self.specgloss_type == self.TEXTURE_FACTOR:
+        elif pyext.specgloss_type == pyext.TEXTURE_FACTOR:
 
-            self.specularGlossinessTexture.blender_create()
+            BlenderTexture.create(pyext.specularGlossinessTexture)
 
             spec_text = node_tree.nodes.new('ShaderNodeTexImage')
-            spec_text.image = bpy.data.images[self.specularGlossinessTexture.image.blender_image_name]
+            spec_text.image = bpy.data.images[pyext.specularGlossinessTexture.image.blender_image_name]
             spec_text.color_space = 'NONE'
             spec_text.location = -1000,0
 
             spec_math     = node_tree.nodes.new('ShaderNodeMath')
             spec_math.operation = 'MULTIPLY'
-            spec_math.inputs[0].default_value = self.glossinessFactor
+            spec_math.inputs[0].default_value = pyext.glossinessFactor
             spec_math.location = -250,100
 
             spec_mapping = node_tree.nodes.new('ShaderNodeMapping')
@@ -347,7 +291,7 @@ class KHR_materials_pbrSpecularGlossiness():
 
             spec_uvmap = node_tree.nodes.new('ShaderNodeUVMap')
             spec_uvmap.location = -1500,0
-            spec_uvmap["gltf2_texcoord"] = self.specularGlossinessTexture.texcoord # Set custom flag to retrieve TexCoord
+            spec_uvmap["gltf2_texcoord"] = pyext.specularGlossinessTexture.texcoord # Set custom flag to retrieve TexCoord
 
 
             # links
@@ -364,15 +308,3 @@ class KHR_materials_pbrSpecularGlossiness():
         node_tree.links.new(mix.inputs[2], diffuse.outputs[0])
         node_tree.links.new(mix.inputs[1], glossy.outputs[0])
         node_tree.links.new(output_node.inputs[0], mix.outputs[0])
-
-
-    def debug_missing(self):
-        if self.json is None:
-            return
-        keys = [
-
-                ]
-
-        for key in self.json.keys():
-            if key not in keys:
-                self.gltf.log.debug("KHR_materials_pbrSpecularGlossiness MISSING " + key)
